+++ js/core/app.js
@@ -1,7 +1,7 @@
 (function () {
   'use strict';
 
-  const APP_VERSION = 'v22';
+  const APP_VERSION = 'v23';
 
   const App = {
     version: APP_VERSION,